requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
pip
pytz
setuptools
sqlparse
tzdata
wheel
Pillow
pydantic
